added parsing from medium RSS xml
